Split into a header and cpp file, modified the rendering.
Now it renders by copying the input channels through to the output, darkening
any area that isn't a face. This still needs a bit of work: it's almost
invisible when used on a dark scene lke that Avatar shot.

Also, it ignores the circlecolor knob. Need to do something about that.