Split the plugin into separate source and header files, and improved the rendering.
The rendering is now done by copying darkening the input channel
wherever we don't find a face. It's now only writing to an IplImage,
never rendering into it or reading back from it. This eliminates a lot
of copying and data transformation as well as giving a better quality
end image (no degradation due to conversion to and from an 8-bits per
channel format).

The downside of this new rendering approach is that it isn't as visible
as the old one for darker images; and that it ignores the circle color
control.

Also updated the MacOS X build file to prevent it building for the PPC
architecture. It seemed a bit wasteful. If you want to switch it back
on, add -arch ppc to both the compile and link flags.